Added doc strings to both the cluster.py and gateway.py
Doc strings describe the allowed parameters and their types along with some usages. One was added for HTCGatewayCluster which is not seen by the user. The new_cluster function in the HTCGateway class was edited to reflect the new kwargs and options usage. The gateway now provides the doc string upon certain errors.